refactor(architecture): thin release command
