+++ src/core/release/mod.rs
@@ -4,10 +4,14 @@ mod pipeline;
 mod resolver;
 mod types;
 mod utils;
+mod workflow;
 pub mod version;
 
 pub use pipeline::{plan, run};
 pub use types::{
-    ReleaseArtifact, ReleaseOptions, ReleasePlan, ReleasePlanStatus, ReleasePlanStep, ReleaseRun,
+    ReleaseArtifact, ReleaseCommandInput, ReleaseCommandResult, ReleaseDeploymentResult,
+    ReleaseDeploymentSummary, ReleaseOptions, ReleasePlan, ReleasePlanStatus, ReleasePlanStep,
+    ReleaseProjectDeployResult, ReleaseRun,
 };
 pub use utils::{extract_latest_notes, parse_release_artifacts};
+pub use workflow::run_command;
